cond-PPM byte_0 mix: proper marginal for include_space (fixes per-token-uniform fallback) — val_bpb 1.015784
