--- records/track_10min_16mb/2026-04-30_OptE_SlidingWindow_CondPPM/eval_seed314.log
@@ -1,185 +1,0 @@
-W0430 22:58:17.050000 596248 torch/distributed/run.py:774] 
-W0430 22:58:17.050000 596248 torch/distributed/run.py:774] *****************************************
-W0430 22:58:17.050000 596248 torch/distributed/run.py:774] Setting OMP_NUM_THREADS environment variable for each process to be 1 in default, to avoid your system being overloaded, please further tune the variable for optimal performance in your application as needed. 
-W0430 22:58:17.050000 596248 torch/distributed/run.py:774] *****************************************
-Hyperparameters:
-  adam_eps: 1e-08
-  adam_wd: 0.02
-  artifact_dir: 
-  attn_clip_sigmas: 13.0
-  attn_out_gate_enabled: False
-  attn_out_gate_src: proj
-  awq_lite_bits: 8
-  awq_lite_enabled: False
-  awq_lite_group_size: 64
-  awq_lite_group_top_k: 1
-  beta1: 0.9
-  beta2: 0.95
-  caseops_enabled: True
-  compressor: brotli
-  datasets_dir: /dev/shm/pgdata/datasets/fineweb10B_sp8192_lossless_caps_caseops_v1_reserved
-  distributed: True
-  ema_decay: 0.9965
-  embed_bits: 6
-  embed_clip_sigmas: 15.0
-  embed_lr: 0.6
-  embed_wd: 0.085
-  enable_looping_at: 0.65
-  eval_seq_len: 2048
-  eval_stride: 64
-  fused_ce_enabled: True
-  gate_window: 12
-  gated_attn_enabled: False
-  gated_attn_init_std: 0.01
-  gated_attn_quant_gate: False
-  global_ttt_batch_seqs: 32
-  global_ttt_chunk_tokens: 32768
-  global_ttt_epochs: 1
-  global_ttt_grad_clip: 1.0
-  global_ttt_lr: 0.001
-  global_ttt_momentum: 0.9
-  global_ttt_respect_doc_boundaries: True
-  global_ttt_warmup_chunks: 0
-  global_ttt_warmup_start_lr: 0.0
-  gptq_calibration_batches: 16
-  gptq_reserve_seconds: 0.5
-  grad_accum_steps: 1
-  grad_clip_norm: 0.3
-  is_main_process: True
-  iterations: 20000
-  jepa_aux_weight: 0.0
-  ln_scale: True
-  local_rank: 0
-  logfile: logs/E_diag_v2_seed314.txt
-  logit_softcap: 30.0
-  loop_end: 5
-  loop_start: 3
-  lqer_asym_enabled: True
-  lqer_asym_group: 64
-  lqer_enabled: True
-  lqer_factor_bits: 4
-  lqer_rank: 4
-  lqer_top_k: 3
-  macaron_enabled: False
-  matrix_bits: 6
-  matrix_clip_sigmas: 11.5
-  matrix_lr: 0.026
-  max_wallclock_seconds: 600.0
-  min_lr: 0.1
-  mlp_clip_sigmas: 11.5
-  mlp_mult: 4.0
-  model_dim: 512
-  model_path: final_model.pt
-  mtp_heads: 0
-  mtp_weight: 0.3
-  multi_exit_aux_weight: 0.1
-  multi_exit_enabled: False
-  multi_exit_layers: 4,6,8
-  multi_exit_mix_lr: 0.05
-  multi_exit_mix_steps: 80
-  muon_backend_steps: 5
-  muon_momentum: 0.97
-  muon_momentum_warmup_start: 0.92
-  muon_momentum_warmup_steps: 1500
-  muon_row_normalize: True
-  muon_wd: 0.095
-  num_heads: 8
-  num_kv_heads: 4
-  num_layers: 11
-  num_loops: 2
-  parallel_final_lane: mean
-  parallel_start_layer: 5
-  phased_ttt_num_phases: 1
-  phased_ttt_prefix_docs: 2000
-  ppm_byte_conditional_alpha: 15.0
-  ppm_byte_conditional_beta: 0.8
-  ppm_byte_conditional_enabled: True
-  ppm_conf_threshold: 0.9
-  ppm_enabled: True
-  ppm_gate_mode: binary
-  ppm_lambda_hi: 0.9
-  ppm_lambda_lo: 0.05
-  ppm_mix_level: byte
-  ppm_order: 5
-  ppm_sigmoid_alpha: 15.0
-  ppm_sigmoid_beta: 0.8
-  ppm_subset_tokens: 5000000
-  ppm_token_conf_aggregate: mean
-  prequant_ttt_batch_seqs: 32
-  prequant_ttt_beta1: 0.9
-  prequant_ttt_beta2: 0.999
-  prequant_ttt_chunk_tokens: 32768
-  prequant_ttt_compile: True
-  prequant_ttt_enabled: False
-  prequant_ttt_epochs: 21
-  prequant_ttt_fedavg_weights: True
-  prequant_ttt_grad_clip: 1.0
-  prequant_ttt_lr: 0.0005
-  prequant_ttt_lr_final: 5e-05
-  prequant_ttt_optimizer: adamw
-  prequant_ttt_weight_decay: 0.0
-  qk_gain_init: 5.0
-  quantized_model_path: final_model.int6.ptz
-  rank: 0
-  rope_base: 10000.0
-  rope_dims: 16
-  rope_train_seq_len: 2048
-  rope_yarn: False
-  run_id: E_diag_v2_seed314
-  scalar_lr: 0.02
-  seed: 314
-  skip_gates_enabled: True
-  sliding_window_batch_seqs: 8
-  sliding_window_enabled: True
-  smear_gate_enabled: True
-  sparse_attn_gate_enabled: True
-  sparse_attn_gate_init_std: 0.0
-  sparse_attn_gate_scale: 1.0
-  stoch_depth_max: 0.02
-  stoch_depth_schedule: linear
-  temp_cal_enabled: False
-  temp_cal_lr: 0.1
-  temp_cal_steps: 50
-  tie_embeddings: True
-  tied_embed_init_std: 0.005
-  tied_embed_lr: 0.03
-  tokenizer_path: ./tokenizers/fineweb_8192_bpe_lossless_caps_caseops_v1_reserved.model
-  train_batch_tokens: 786432
-  train_files: /dev/shm/pgdata/datasets/fineweb10B_sp8192_lossless_caps_caseops_v1_reserved/fineweb_train_*.bin
-  train_log_every: 500
-  train_seq_len: 2048
-  ttt_batch_size: 64
-  ttt_beta1: 0.0
-  ttt_beta2: 0.999
-  ttt_chunk_size: 48
-  ttt_enabled: False
-  ttt_eval_batches: 
-  ttt_eval_seq_len: 2048
-  ttt_grad_steps: 1
-  ttt_k_lora: True
-  ttt_lora_lr: 0.0001
-  ttt_lora_rank: 96
-  ttt_mlp_lora: True
-  ttt_o_lora: True
-  ttt_optimizer: adam
-  ttt_weight_decay: 1.0
-  val_batch_tokens: 524288
-  val_bytes_files: /dev/shm/pgdata/datasets/fineweb10B_sp8192_lossless_caps_caseops_v1_reserved/fineweb_val_bytes_*.bin
-  val_doc_fraction: 1.0
-  val_files: /dev/shm/pgdata/datasets/fineweb10B_sp8192_lossless_caps_caseops_v1_reserved/fineweb_val_*.bin
-  val_loss_every: 4000
-  vocab_size: 8192
-  warmdown_frac: 0.85
-  warmup_steps: 20
-  world_size: 8
-  xsa_last_n: 11
-train_shards: 97
-val_tokens: 9662464
-TTT_EVAL_ONLY=1 — skipping training + GPTQ, loading saved artifact for TTT eval
-ttt_lora_alpha: 144.0
-ttt_warm_start_a: True
-ttt_weight_decay: 1.0
-diagnostic quantized val_loss:2.53463309 val_bpb:1.17928735 eval_time:7577ms
-cond_ppm tokens=9662464 bytes=32756252 cond_mix_bpb=1.026967 alpha=15.0 beta=0.8
-quantized_cond_ppm val_loss:2.41317110 val_bpb:1.02696665
-quantized_sliding_window val_loss:2.54444884 val_bpb:1.18426300 eval_time:227686ms
